Updates default og image and theme
Updates the default og image to og.png.

Removes unused images from config.

Switches default theme to dark.

Updates package version.
