+++ gl-core/config.json
@@ -8,9 +8,8 @@
     "favicon": "/png/favicon.png"
   },
   "imagesArr": [
-    "/png/favicon.png",
-    "/png/default.png",
-    "/png/bluesky/javascript.png"
+    "/png/apple-touch-icon.png",
+    "/png/og.png"
   ],
   "themes": {
     "light": {
