+++ package.json
@@ -1,5 +1,5 @@
 {
-  "version": "51.5.1",
+  "version": "51.5.2",
   "name": "core",
   "description": "Always FREE and Open Source",
   "private": true,
